+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.2.37
+
+### Patch Changes
+
+-   36518cc: check cloud systems are connected before making backend call
+
 ## 1.2.36
 
 ### Patch Changes