+++ packages/cf-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.1.29
+
+### Patch Changes
+
+-   37c2e00: remove fs, causing issues when cds is updating package.json
+
+## 0.1.28
+
+### Patch Changes
+
+-   a8cbdcf: append install script
+
 ## 0.1.27
 
 ### Patch Changes